+++ CHANGELOG.md
@@ -1419,3 +1419,4 @@
 - [2026-05-24 19:24] docs(perf): update readme and API contracts for perf
 - [2026-05-24 19:57] feat(i18n): add async i18n stream provider
 - [2026-05-24 20:23] feat(analytics): add bi-directional event bus handler for analytics
+- [2026-05-24 21:10] fix(i18n): handle null assertion failure in i18n parser